Put $Id$ on all of these.
Correct Makefile so that we build during all: and only install for
make install.

If /etc/sysconfig exists source it to get the flags for restarting
named with.  If /etc/sysconfig says no named runs, don't try to start one.

Don't attempt to kill anything if we can not find the named.pid file.

Reviewed by:	Nickolay N. Dudorov <nnd@gw.itfs.nsk.su>

+++ usr.sbin/named/tools/named.restart/named.restart.sh
@@ -1,12 +1,23 @@
 #!/bin/sh -
 #
 #	from named.restart	5.4 (Berkeley) 6/27/89
-#	$Id: named.restart.sh,v 4.9.1.5 1994/04/09 03:43:17 vixie Exp $
+#	$Id: named.restart.sh,v 1.2 1994/09/22 20:45:26 pst Exp $
 #
+
+# If there is a global system configuration file, suck it in.
+if [ -f /etc/sysconfig ]; then
+	. /etc/sysconfig
+fi
 
 PATH=%DESTSBIN%:/bin:/usr/bin
 
-pid=`cat %PIDDIR%/named.pid`
-kill $pid
-sleep 5
-exec %INDOT%named 
+if [ -f %PIDDIR%/named.pid ]; then
+	pid=`cat %PIDDIR%/named.pid`
+	kill $pid
+	sleep 5
+fi
+
+# $namedflags is imported from /etc/sysconfig
+if [ "X${namedflags}" != "XNO" ]; then 
+	exec %INDOT%named ${namedflags}
+fi
